Updating to 1.7.10, WIP

+++ src/net/bdew/deepcore/multiblock/interact/CIPowerProducer.scala
@@ -10,5 +10,5 @@
 package net.bdew.deepcore.multiblock.interact
 
 trait CIPowerProducer extends CIOutputFaces {
-  def extract(v: Float, simulate: Boolean): Float
+  def extract(v: Double, simulate: Boolean): Float
 }
